requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
django-redis==5.0.0
djangorestframework==3.12.4
idna==3.2
install==1.3.4
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
